# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NyayaSetu Legal Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # DASHBOARD
...
  /rights:
    get:
      operationId: listRights
      tags: [rights]
      summary: List legal rights articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Rights articles list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RightsListResponse"
...
  /rights/{id}:
    get:
      operationId: getRightsArticle
      tags: [rights]
      summary: Get a single rights article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rights article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RightsArticle"
        "404":
          description: Not found

  # COMPLAINTS
...
  /schemes:
    get:
      operationId: listSchemes
      tags: [schemes]
      summary: List government schemes
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Schemes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemesListResponse"
  /schemes/{id}:
    get:
      operationId: getScheme
      tags: [schemes]
      summary: Get scheme details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scheme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scheme"

  # LEGAL LEARNING HUB
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationReadInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /notifications/mark-all-read:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyInput"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # GLOBAL SEARCH
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    EmptyInput:
      type: object
      properties: {}
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        aadhaar:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        preferredLanguage:
          type: string
        profilePhoto:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
      required: [id, fullName, email, preferredLanguage, role, createdAt]
...
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        mobile:
          type: string
        aadhaar:
          type: string
        address:
          type: string
        preferredLanguage:
          type: string
        profilePhoto:
          type: string
    PasswordChangeInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
...
    RightsArticle:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        category:
          type: string
        relevantLaws:
          type: array
          items:
            type: string
        legalProvisions:
          type: array
          items:
            type: string
        governmentResources:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, title, summary, content, category, relevantLaws, legalProvisions, governmentResources, createdAt]
    RightsListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/RightsArticle"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [articles, total, page, totalPages]
...
    Scheme:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        eligibility:
          type: string
        benefits:
          type: string
        requiredDocuments:
          type: array
          items:
            type: string
        applicationProcess:
          type: string
        officialLink:
          type: ["string", "null"]
        ministry:
          type: string
        createdAt:
          type: string
      required: [id, title, description, category, eligibility, benefits, requiredDocuments, applicationProcess, ministry, createdAt]
    SchemesListResponse:
      type: object
      properties:
        schemes:
          type: array
          items:
            $ref: "#/components/schemas/Scheme"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [schemes, total, page, totalPages]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, type, title, message, isRead, createdAt]
    NotificationReadInput:
      type: object
      properties:
        isRead:
          type: boolean
      required: [isRead]
...