# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NyayaSetu Legal Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordChangeInput {
  currentPassword: string;
  newPassword: string;
}